[MRG] split ot.bregman  (#569)
* draft bregman split

* update contrib+release+check tests

* fix tests

* fix?

* fix docstring issues

* add missing contributors
