+++ RELEASES.md
@@ -23,6 +23,8 @@
 - Handle documentation and warnings when integers are provided to (f)gw solvers based on cg (Issue #530, PR #559)
 - Correct independence of `fgw_barycenters` to `init_C` and `init_X` (Issue #547, PR #566)
 - Avoid precision change when computing norm using PyTorch backend (Discussion #570, PR #572)
+- Create `ot/bregman/`repository (Issue #567, PR #569)
+
 
 ## 0.9.1
 *August 2023*
